Change nmaster from uint16_t to int and fix compilation problem in \!DAGUE_GPU_CUDA_ALLOC_PER_TILE mode

+++ include/moesi.h
@@ -32,7 +32,7 @@ typedef uint32_t moesi_key_t;
  * Blocks that are not locally stored have a NULL master
  */
 struct _moesi_map {
-    uint16_t nmasters;
+    int      nmasters;
     uint16_t ndevices;
     moesi_master_t* masters[1];
 };
